Support the import of existing TPM-wrapped keys.
This CL makes it possible to create a private key object which already
holds a TPM-wrapped private key using the conventional Chaps vendor
attributes.  This allows keys used in other parts of the system to be
imported into a PKCS #11 token.

Previously the read-only-on-create policy was set to true for internal
attributes because there was no reason not to.  Now that we have a use
case for creating objects with internal attributes, this restriction has
been lifted.

BUG=chromium-os:37815
TEST=Unit tests; Manually tested importing an existing key.

Change-Id: I457da852b999d323dbdd6cb39b3b1f37f550fa1b
Reviewed-on: https://gerrit.chromium.org/gerrit/45534
Reviewed-by: Gaurav Shah <gauravsh@chromium.org>
Commit-Queue: Darren Krahn <dkrahn@chromium.org>
Tested-by: Darren Krahn <dkrahn@chromium.org>