Salsa: Use radio buttons for 2-treatment exps
For experiments with only two treatments, it's silly to have the
additional complecation of a stack-rank mechanism.  There are only three
options for them:
 * Treatment 0 is better
 * Treatment 1 is better
 * They are the same
So for the sake of simplicity, just using radio buttons to pick between
them is better.  The users already know how they work and don't have to
think about it.

This patch checks how many treatments there are when the participant
goes to the page and either serves them the full stack-rank system or
the simplified radio-button ranking.

BUG=chromium:187154
TEST=manually run using a dev server

Change-Id: Ibc499fa107632256e65ae0a859aaf675439e7e4a
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>