Enforce the StartUpFlags device policy through session manager.
In order to make sure flags specified through policy are enforced
and applied before the browser had the chance to initialize we
should set them through session manager before it spawns the
browser process. This feature will not be exposed until we have
some way to ensure that only the set of flags inteded to be used
through this mechanism is enfoced either in the session manager or
on the browser side.

BUG=chromium-os:39251
TEST=Set the policy and check that the policy is respected.

Change-Id: I1bc5e0bfd93bcb61230a0c87f9d63edee00dfaea
Reviewed-on: https://gerrit.chromium.org/gerrit/45434
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Julian Pastarmov <pastarmovj@chromium.org>
Commit-Queue: Julian Pastarmov <pastarmovj@chromium.org>