ipsec-manager: Accept "tunnel group" parameter
Specify to "Agressive Mode" and encode the tunnel group name
when asked to do so via a command-line flag.  This allows a
user realm to be specified during the first round of negotiations,
which is interpreted by Cisco VPNs as the Tunnel Group for which
various configuration and policy is layered.

BUG=chromium:199004
TEST=Use in tandem with https://gerrit.chromium.org/gerrit/46154
and the new "--tunnel-group" parameter to the "connect-vpn" test
script and a Cisco ASA 5505 VPN configured with an alterate tunnel
group.  Unfortunately due to configuration issues on the VPN, the
IPSec connection was observed to be established, and the correct
tunnel group was indicated in the logs, but the PPP link did not
come up due to an AAA internal issue to the VPN.  Verify that
connections to the default tunnel group continue to work correctly
without the "--tunnel-group" flag.

Change-Id: Ie53bbe8dd1c16a72ae2265d4879e5435fb23d73e
Reviewed-on: https://gerrit.chromium.org/gerrit/46153
Commit-Queue: Paul Stewart <pstew@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>