service.cc: propagate enterprise ownership to Mount instances
In the past, there was one Mount that was reused for many operations
across cryptohomed.  This didn't scale for multiple simultaneous
mounts and was factored out.  The move to a MountFactory and per-cryptohome
Mount instances left enterprise ownership unannotated.  This change
brings it back by populating when the Mount is created or after the
fact if finalization follows later.

Signed-off-by: Will Drewry <wad@chromium.org>
TEST=builds, unittests pass, (testing this one now) runs on lumpy and the cryptohome status shows enterprise enrollment
BUG=chromium:196621

Change-Id: I23e57d58a3d66a89296bfc9098afa87f197f77cc
Reviewed-on: https://gerrit.chromium.org/gerrit/46267
Reviewed-by: Bartosz Fabianowski <bartfab@chromium.org>
Commit-Queue: Will Drewry <wad@chromium.org>
Tested-by: Will Drewry <wad@chromium.org>