shill: Notify Manager on Service.AutoConnect property changes.
This allows shill to take into account property changes and, if
necessary, initiate auto-connect immediately after the property is
set.

BUG=chromium:223544
TEST=unit tests; configured a VPN service, then marked it as
auto-connect -- observed in the logs that shill suppressed an
auto-connect to it right away due to "never-connected"; configured a
VPN service, connected successfully, disconnected, then marked it as
auto-connect -- observed shill trying to auto-connect to the service.

Change-Id: Ie3a0950201245dccd9662c15f76ef0e5b34e9d67
Reviewed-on: https://gerrit.chromium.org/gerrit/46398
Reviewed-by: Darin Petkov <petkov@chromium.org>
Tested-by: Darin Petkov <petkov@chromium.org>
Reviewed-by: Philipp Neubeck <pneubeck@chromium.org>
Commit-Queue: Darin Petkov <petkov@chromium.org>