Enable zram by default.
This change enables zram by default for all platforms.
It still allows manual control of zram from crosh.

Before this change, zram was not enabled if the file
/home/chronos/.swap_enabled was missing.  Now it is
enabled with default swap size if the file is missing
or empty.  It is disabled if the file indicates a
size of "0".  Otherwise the content of the file
specifies the size.

The default swap size is 1.5 x RAM size, but it
can be overridden by /etc/default/zram_size.sh.

Before this change goes in, we also need to either end
the zram field trial, or change it to behave correctly
with the new defaults.

BUG=chromium:218338
TEST=manually tested

Change-Id: If6a208e317864fb25e1e6181f57910d6b8e2593e
Reviewed-on: https://gerrit.chromium.org/gerrit/46402
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Queue: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>