Crosh: compressed swap ON by default
These changes are needed when compressed swap is turned on
by default in /etc/init/swap.conf.  These are mostly changes
in messages to reflect the new situation, and a change in
the way swap is disabled.

We wish to leave this hook for a little longer, mostly to
be able to keep experimenting, and to turn swap off if it
proves to be a disaster in specific circumstances (i.e.
under unexpected loads).

BUG=chromium:218338
TEST=manual

Change-Id: Ic0a2d61c3d930a8bc163e69dd998eddc5f32f10d
Reviewed-on: https://gerrit.chromium.org/gerrit/46417
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Queue: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>