power: Handle in-progress async I/O operations.
This updates AsyncFileReader::Reset() to call aio_suspend()
to wait for the in-progress request to complete when
aio_cancel() returns AIO_NOTCANCELED to indicate that the
request was in-progress and couldn't be canceled.
Otherwise, the subsequent completion of the request can
cause a crash.

BUG=chromium:220740
TEST=powerd behaves the same as before.  also ran
     system_unittest in a loop with
     --gtest_filter='AsyncFileReaderTest.*:AmbientLightSensorTest.*'
     and observed that it didn't crash when the
     AIO_NOTCANCELED condition is hit

Change-Id: I58ba14aba87ae8af2cdafe9d964d8e33a739e376
Reviewed-on: https://gerrit.chromium.org/gerrit/46968
Tested-by: Daniel Erat <derat@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Commit-Queue: Daniel Erat <derat@chromium.org>