power: Force backlight on for user activity.
This ensures that the panel backlight is turned on if the
user previously sets the brightness to zero but later begins
interacting with the computer again (but only if an external
display isn't connected).

Note that decreasing the brightness to zero and then
immediately hitting a key typically won't turn the backlight
back on; this is due to Chrome rate-limiting its reports of
user activity such that at most one is sent every five
seconds.  To avoid this, Chrome would probably need to
report every user action to powerd.

BUG=chromium:221303
TEST=added unit test; also manually tested

Change-Id: I48e2c29178edeb6b5f4af84b609edfa864f0dc35
Reviewed-on: https://gerrit.chromium.org/gerrit/46969
Tested-by: Daniel Erat <derat@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Daniel Erat <derat@chromium.org>