shill: Add ConnectionHealthChecker class to check connectivity status.
ConnectionHealthChecker asynchronously creates a TCP connection with a
remote server and checks that either (1) a small amount of data can be
sent over the connection, or (2) the connection can be cleanly shut
down. ConnectionHealthChecker may be provided multiple remote server
addresses, and it attempts them in a round-robin fashion if one attempt
fails.

BUG=chromium:225913
TEST=Build and run unit tests.

Change-Id: I01013af2214fd5614b1fca820670de12757a8904
Reviewed-on: https://gerrit.chromium.org/gerrit/47752
Commit-Queue: Ben Chan <benchan@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>