shill: Break circular reference between MockManager and EthernetService
The Ethernet unit tests crash currently crash when built against
gmock/gtest-1.6.

The expectation EXPECT_CALL(manager_, DeregisterService(GetService())) creates
a copy of the return value of GetService(), a ServiceRefPtr. The mock manager
will hold onto this reference until its destruction.

Trying to destroy a service while the manager is being destroyed is problematic
given the reference relationships. The ethernet service holds a reference to
the Ethernet device, whose destructor depends on the manager still being alive.

So, re-write the expectation such that only a local reference is held which
ensures that service and device are already destroyed by the time the mock
manager destructor is called.

BUG=chromium:211445
TEST=build and run shill unit tests

Change-Id: I6f9463131468f7be5fe600194285dc66c523ba53
Reviewed-on: https://gerrit.chromium.org/gerrit/47618
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Gaurav Shah <gauravsh@chromium.org>
Tested-by: Gaurav Shah <gauravsh@chromium.org>