Separate file access from perf data read/write
Read:
Instead of reading each section of perf data from the file and storing
it, read all data from the file into a buffer, and then read from the
buffer.

Write:
Instead of writing each part of perf data to the file as it is
generated, write them to a buffer, and then write the buffer to a file
in one shot.

This decouples the perf data reading from file access.

BUG=none
TEST=unit tests pass

Change-Id: I2c42a6f469ee906c826989ac53578d5a95e2b603
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/48130
Reviewed-by: asharif <asharif@chromium.org>