power: Make AmbientLightHandler power-source aware.
This updates the AmbientLightHandler class to support
returning different brightness levels when on AC power vs.
battery power.  It's likely that we need this for the
internal panel's backlight.

A separate target percentage for battery power can be
specified in the second field of a line in the
internal_backlight_als_steps or keyboard_backlight_als_steps
prefs.  If only three fields are supplied, the percentage in
the first field is used for both AC and battery.

BUG=chromium:188826
TEST=added unit tests for AmbientLightHandler (it was
     previously only covered by KeyboardBacklightController
     and InternalBacklightController's tests); manually
     tested with a custom internal_backlight_als_steps pref

Change-Id: I4c2c761f70ab2b257d1491cc122ce757f88bc0a4
Reviewed-on: https://gerrit.chromium.org/gerrit/48084
Tested-by: Daniel Erat <derat@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Commit-Queue: Daniel Erat <derat@chromium.org>