power: Update displayed battery percentages.
This simplifies the previous "pin to 100% for three minutes,
then taper down to the actual percentage over seven minutes"
behavior when going from AC to battery with a full charge.
Instead, the charge at the point when the battery was
considered full is used as the high-water mark for the
entirety of the discharge cycle.  This avoids
discontinuities in the displayed percentage when connecting
and disconnecting AC power.

It also makes an improvement to a logging message that I
intended to include in a previous patch. :-/

I'm also sneaking in a small check to avoid negative
time-to-full durations if the battery reports a higher
charge than its supposedly maximum charge, which I appear to
have just seen on a Pixel PVT device.

BUG=chromium:230112
TEST=updated tests; manually tested that levels are
     continuous when switching between AC and battery power
     near 100%

Change-Id: Ic06e6aea4eaf5a80df43aef527e3139601914eaf
Reviewed-on: https://gerrit.chromium.org/gerrit/48914
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Commit-Queue: Daniel Erat <derat@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>