shill: WiFiService: Expose Country Code of AP
Report the 2-letter country code associated with the representative
endpoint of a WiFi service by extracting the information element in
the Endpoint and displaying this as a read-only property in the
Service.

BUG=chrome-os-partner:18659
TEST=Unit test + Manual: list-services should display a "Country"
property for services.  APs that don't report a country code will
have an empty Country property.

Change-Id: I495427860fe57a0cfa999d701a8fd29ed4c4c4f5
Reviewed-on: https://gerrit.chromium.org/gerrit/48366
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>