shill_login_user: Use InsertUserProfile instead
Use a different call handle to add user profiles which allows
the user hash to be passed.  An empty user hash is sent if
this is the guest user.

CQ-DEPEND=CL:48603
BUG=chromium:231858
TEST=Manual: Login and ensure that /var/run/shill/loaded_profile_list
contains not just the username but the hash.  Use "list-profiles" to
confirm that the "UserHash" property is set on the user profile.
Restart shill and make sure the "UserHash" property is still set on
the profile in "list-services".  Logout and login as guest.  Ensure
that "list-profiles" shows a user profile but one that contains an
empty "UserHash" property.  Restart shill to ensure that the profile
gets reloaded.  Modified version of network_ShillInitScripts

Change-Id: I9da8011f7bcb6ffab63a7c674d92c546ef2a573d
Reviewed-on: https://gerrit.chromium.org/gerrit/48584
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>