shill: EapCredentials: Never reset key_management_
The "key_management_" field of the EapCredentials is special in
that it is set in WiFi during the constructor and in many flows
it may not ever be set again.  It is required to be non-empty
during connection to an AP.  There are already safeguards in
EapCredentials::SetKeyManagement().  Fix EapCredentials::Reset()
to avoid resetting it here as well.

BUG=chromium:234396
TEST=New unit test

Change-Id: I23ec552ea3fad41372bf91a02c14f68618a7411c
Reviewed-on: https://gerrit.chromium.org/gerrit/48950
Commit-Queue: Paul Stewart <pstew@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>