chaps: Provide a pam module which logs in a user to chaps on Linux.
This change adds a pam module to chaps which is built when building chaps
for Linux. When this module is added, it will open an isolate for the user
when they login, and load their user token into that isolate.  On logout,
the isolate will be closed, unloading the token if all sessions have
been closed.

The change also adds a simple "make install" option for Linux.

TEST=Built and run on Linux, testing pam module login / logout /
change password functionality.
BUG=None
CQ-DEPEND=CL:50461

Change-Id: I0cf443873c895e1e2cd6ea966a4ab26caa2a9f34
Reviewed-on: https://gerrit.chromium.org/gerrit/49331
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Darren Krahn <dkrahn@chromium.org>
Commit-Queue: Ross McIlroy <rmcilroy@chromium.org>
Tested-by: Ross McIlroy <rmcilroy@chromium.org>