chaps: Change chaps dbus config to have configurable policy permissions.
This change enables the chaps dbus configuration to be configured depending
upon the platform it is being installed on. On ChromeOS it will continue to
be only accessable to members of the pkcs11 group. On Linux, it will be
accessable to any user, since on Linux we use isolates to ensure a user can
only access their own token.

TEST=Built and ran tests. Tested using trybot. Ran on Linux to ensure correct
dbus interaction.
BUG=None
CQ-DEPEND=CL:50460

Change-Id: I02f90d386ca3c851be93858410af60ca3b4e14a0
Reviewed-on: https://gerrit.chromium.org/gerrit/49594
Tested-by: Ross McIlroy <rmcilroy@chromium.org>
Reviewed-by: Darren Krahn <dkrahn@chromium.org>
Commit-Queue: Ross McIlroy <rmcilroy@chromium.org>