chaps: Add libmemenv check to chaps build, and disable memenv if not available.
The libmemenv library is not part of the standard libleveldb-dev debian
package. We only use this library for testing, so to enable chaps to be
built as a debian package, add a check for whether libmemenv is available
when building, and disable it's usage if it is not available.

TEST=built and ran tests on Linux and cbuildbot
BUG=None

Change-Id: I19594ceccb2c82a87fa8938ec93f85e92d42a9cc
Reviewed-on: https://gerrit.chromium.org/gerrit/49597
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Darren Krahn <dkrahn@chromium.org>
Tested-by: Ross McIlroy <rmcilroy@chromium.org>
Commit-Queue: Ross McIlroy <rmcilroy@chromium.org>