shill: fix hidden networks in default profiles
When shill starts up, Manager::InitializeProfiles restores the
profile stack. This includes the default profile, as well as any
necessary user profiles. (User profiles are necessary if shill
aborted with user profiles on the stack.)

In the process of restoring the profiles, Manager normally asks
each technology Provider to examine the profile for Services
that need to be instantiated. Of particular relevance to this
bug is that WiFiProvider instantiates hidden WiFiServices from
the profile(s).

In order to automatically connect to hidden networks that are
stored in the default profile, Manager:InitializeProfiles needs
to hand the default profile to the WiFiProvider. This didn't
happen before, because the default profile was special-cased.

Fix this by using PushProfileInternal for the default profile,
just like we do for user profiles.

BUG=chromium:221021
TEST=new unit test, manual

Manual test
-----------
- log in
- # stop shill
- # rm /var/cache/shill/default.profile
- # start shill
- log out
- while logged out, connect to a hidden network
- reboot
- observe that shill reconnects to the hidden network

Change-Id: I5b0e517d003522cc8da63d5502cad173ef1afe04
Reviewed-on: https://gerrit.chromium.org/gerrit/49838
Commit-Queue: mukesh agrawal <quiche@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>