AddressMapper stores mapping IDs instead of names
Store a uint64 identifier with each mapping (default is kuint64max),
instead of a name string.  This keeps the mapping data smaller and
simpler.

Update PerfParser to use the new AddressMapper API.  The ID is the mmap
event's index within the set of all perf events.  By using this index
instead of a name string to identify the mappings, separate mappings
with the same DSO name can be uniquely identified.

BUG=chromium:237399
TEST=unit tests pass

Change-Id: I9b0e2947e74897da55a1511a43832fbe7375afde
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/49918
Reviewed-by: asharif <asharif@chromium.org>